niri-config: (WIP) sketch UX for more flexible input configuration
This is an example for what more flexible input configuration could look like in
Niri.

In this model, the inputs section has a set of match blocks which are matched
and applied in order. Matches are selected by:

* match type (required), corresponding to the current types (keyboard, ...)
* match name (optional), match on the human-readable device name
  Idea is that this should be a substring match.
* match product (optional), match on the numeric product ID
* match vendor (optional), match on the numeric vendor ID

The InputRule struct which represents this contains all the possible config
options, so there is no longer any grouping per device type.

This commit is ONLY for testing the syntax and how to map a more flexible
structure to KDL, and what the UX of that looks like.

Things not done yet / open questions:

* actually wire this config up to anything
* decide what to do with nonsensical match type / option combinations